moved upload constants to constants/upload
