+++ src/services/upload/multiPartUploadService.ts
@@ -1,7 +1,10 @@
-import { FILE_CHUNKS_COMBINED_FOR_A_UPLOAD_PART } from './uploadService';
+import {
+    FILE_CHUNKS_COMBINED_FOR_A_UPLOAD_PART,
+    RANDOM_PERCENTAGE_PROGRESS_FOR_PUT,
+} from 'constants/upload';
+import UIService from './uiService';
 import UploadHttpClient from './uploadHttpClient';
 import * as convert from 'xml-js';
-import UIService, { RANDOM_PERCENTAGE_PROGRESS_FOR_PUT } from './uiService';
 import { CustomError } from 'utils/common/errorUtil';
 import { DataStream } from 'types/upload';
